requirements.txt
aiogram==3.4.1
Pillow==10.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.2
openai==1.6.1

# Дополнительные инструменты с GitHub
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
validators==0.22.0
